add training example and remove example model
