--- pytorch/example_model/MNIST/ModelClass.py
@@ -1,22 +1,0 @@
-import torch
-from torch.nn import functional as F
-from torch import nn
-import pytorch_lightning as pl
-
-class PyTorchModel(pl.LightningModule):
-    def __init__(self):
-        super().__init__()
-
-        self.layer_1 = nn.Linear(28 * 28, 128)
-        self.layer_2 = nn.Linear(128, 256)
-        self.layer_3 = nn.Linear(256, 10)
-
-    def forward(self, x):
-        x = self.layer_1(x)
-        x = F.relu(x)
-        x = self.layer_2(x)
-        x = F.relu(x)
-        x = self.layer_3(x)
-
-        x = F.softmax(x, dim=1)
-        return x
