+++ pytorch/pytorch/requirements.txt
@@ -2,6 +2,6 @@ seldon-core==1.6.0
 numpy==1.19.2
 torch==1.8.0
 torchvision==0.9.0
-pytorch_lightning==1.2.5
-cloudpickle==1.4.1
-mlflow==1.17.0
+pytorch_lightning==1.3.0 # required for loading mlflow autologging model
+cloudpickle==1.4.1 # required for loading mlflow autologging model
+mlflow==1.17.0 # required for loading mlflow autologging model
